+++ java-security/src/main/java/com/sap/cloud/security/adapter/xs/XSUserInfoAdapter.java
@@ -5,6 +5,7 @@
  */
 package com.sap.cloud.security.adapter.xs;
 
+import com.sap.cloud.security.client.ClientCredentials;
 import com.sap.cloud.security.config.Environments;
 import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
 import com.sap.cloud.security.config.cf.CFConstants;
