+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/client/OAuth2TokenService.java
@@ -5,6 +5,7 @@
  */
 package com.sap.cloud.security.xsuaa.client;
 
+import com.sap.cloud.security.client.ClientCredentials;
 import com.sap.xsa.security.container.ClientIdentity;
 
 import javax.annotation.Nonnull;
