+++ token-client/src/test/java/com/sap/cloud/security/xsuaa/client/ClientCredentialsTest.java
@@ -5,6 +5,7 @@
  */
 package com.sap.cloud.security.xsuaa.client;
 
+import com.sap.cloud.security.client.ClientCredentials;
 import com.sap.xsa.security.container.ClientIdentity;
 import org.junit.Test;
