+++ token-client/src/test/java/com/sap/cloud/security/xsuaa/client/XsuaaOAuth2TokenServiceRefreshTokenTest.java
@@ -21,6 +21,7 @@ import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 
+import com.sap.cloud.security.client.ClientCredentials;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
